+++ Lab3/TvSet/TvSet.h
@@ -16,15 +16,15 @@ public:
 	bool SelectChannel(string name);
 	bool SelectPreviousChannel();
 	bool SetChannelName(int channelNumber, string channelName);
-	int GetCurrChannel()const;
+	size_t GetCurrChannel()const;
 	bool DeleteChannelName(string name);
 	
 private:
 	typedef map<string, size_t> MapName;
 	typedef map<size_t, string> MapNumber;
 	bool m_isOn = false;
-	int m_currChannel = 1;
-	int m_prevChannel = 1;
+	size_t m_currChannel = 1;
+	size_t m_prevChannel = 1;
 
 	MapName channelName;
 	MapNumber channelNumber;
